+++ appmap/__init__.py
@@ -1,8 +1,10 @@
 """AppMap recorder for Python"""
 
-from ._implementation.env import Env # noqa; F401
+
+from ._implementation.env import Env  # noqa: F401
 from ._implementation.recording import Recording, instrument_module  # noqa: F401
 from ._implementation import generation  # noqa: F401
+from ._implementation.labels import labels  # noqa: F401
 
 
 def enabled():