+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- The `appmap.labels` decorator can now be applied to a function to specify labels that
+  should appear in the AppMap.
+  
+### Fixed
+- [#61] Don't modify an instrumented function's parameters when rendering them.
+- Correct the structure of the `return_value` object in a `return` event.
 
 ## [0.5.0] - 2021-03-08
 ### Added